test: move all old tests from newfeaturetest to ers package after the release
